Moving all the data related files under data_handler
