+++ .env.template
@@ -1,1 +1,4 @@
 OPENAI_API_KEY="YOUR OPENAI KEY"
+PINECONE_API_KEY = "YOUR PINECONE KEY"
+PINECONE_ENVIRONMENT = "PINECONE_ENVIRONMENT"
+GOOGLE_APPLICATION_CREDENTIALS = "PATH TO APPLICATION CREDENTIALS JSON"
